Adicionando infinite scrool na apresentacao de produtos, testes de CRUD do sistema e loading skeleton na paginas
test modificaded

test product CRUD works

test category CRUD works

test add varios products modal works

edit product

button pricipal hover effetc changed

delete category archives in the file works

skeleton and lazy loading add to UI

add skeleton

skeleton loading pages works

testando push para production

test deploy

test deploy

test deploy

test deploy

test deploy

test deploy

test deploy

test deploy

deploying 2

Delete .github/workflows directory

Delete env prod.env

test build

test deployment

test deployment

test deployment

test

teste

teste

teste

teste

teste

teste

teste

teste

teste

teste

teste

fix: guest

test building to deploy

test building to deploy 2

test building to deploy 3

test

build and deploy

deployment

deployment

deployment

deployment

deployment

deployment

deployment

test

add to timing in yml archive

Delete .env.ci

deployment via yml

deployment via yml

deployment via yml

deployment via yml

deployment via yml

deployment via yml

deployment via yml

remove xxxx test to the guest

remove xxxx test to the guest

remove xxxx test to the guest

remove xxxx test to the guest

public storage_path changed

public storage_path changed to public/storage

filesystems link public_path images

filesystems link public_path only /

filesystems link public_path only app

storage don't works

storage don't works

filesystems

filesystems

+++ app/Livewire/ModalConfirm.php
@@ -4,6 +4,7 @@ namespace App\Livewire;
 
 use App\Models\Category;
 use App\Models\Product;
+use Illuminate\Support\Facades\Storage;
 use Livewire\Attributes\On;
 use Livewire\Component;
 
@@ -30,16 +31,44 @@ class ModalConfirm extends Component
     }
 
     public function destroyCategory($id){
-        Category::where('id', $id)->delete();
-        Product::where('category', $id)->delete();
+        $products = Product::where('id_category', $id)->get();
         
+        foreach($products as $product){
+            $image_path = $product->image;
+
+            $image_exists = Storage::disk('public')->exists($image_path);
+
+            if($image_exists){
+                Storage::disk('public')->delete($image_path);
+            }    
+        }
+
+        Product::where('id', $id)->delete();
+        Category::where('id', $id)->delete();
+
         return redirect()->route('view.category')->with('success', 'Categoria deletada com sucesso');   
     }
 
     public function destroyProduct($id){
-        Product::where('id', $id)->delete();
+        $product = Product::where('id', $id)->first();
+
+        if($product){
+            $image_path = $product->image;
+
+            $image_exists = Storage::disk('public')->exists($image_path);
+
+            if($image_exists){
+                Storage::disk('public')->delete($image_path);
+            }
+
+            Product::where('id', $id)->delete();
+            
+            session()->flash('success', 'Produto deletado com sucesso!');
+
+            return redirect()->route('view.products');
+        }
         
-        session()->flash('success', 'Produto deletado com sucesso!');
+        session()->flash('error', 'Nenhum produto encontrado!');
 
         return redirect()->route('view.products');
     }
